+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/github_app_events/GithubAppEvent.java
@@ -1,0 +1,10 @@
+package com.onlydust.marketplace.indexer.domain.models.clean.github_app_events;
+
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@AllArgsConstructor
+@Slf4j
+public class GithubAppEvent {
+
+}
