+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/clean/github_app_events/InstallationEvent.java
@@ -1,8 +1,8 @@
-package com.onlydust.marketplace.indexer.domain.models.clean.events;
+package com.onlydust.marketplace.indexer.domain.models.clean.github_app_events;
 
 import com.onlydust.marketplace.indexer.domain.models.clean.CleanAccount;
 import com.onlydust.marketplace.indexer.domain.models.clean.CleanRepo;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawInstallationEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawInstallationEvent;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
