+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/exposition/GithubAppInstallation.java
@@ -1,6 +1,6 @@
 package com.onlydust.marketplace.indexer.domain.models.exposition;
 
-import com.onlydust.marketplace.indexer.domain.models.clean.events.InstallationEvent;
+import com.onlydust.marketplace.indexer.domain.models.clean.github_app_events.InstallationEvent;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
