+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/RawInstallation.java
@@ -1,7 +1,7 @@
 package com.onlydust.marketplace.indexer.domain.models.raw;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.onlydust.marketplace.indexer.domain.models.clean.events.InstallationEvent;
+import com.onlydust.marketplace.indexer.domain.models.clean.github_app_events.InstallationEvent;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
