--- domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/RawIssueCommentEvent.java
@@ -1,19 +1,0 @@
-package com.onlydust.marketplace.indexer.domain.models.raw;
-
-
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.Value;
-
-@Value
-@EqualsAndHashCode(callSuper = true)
-@NoArgsConstructor(force = true)
-@ToString(callSuper = true)
-public class RawIssueCommentEvent extends JsonDocument {
-    String action;
-    RawIssue issue;
-    RawComment comment;
-    RawRepo repository;
-    RawInstallation installation;
-}
