+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/github_app_events/RawInstallationEvent.java
@@ -1,7 +1,10 @@
-package com.onlydust.marketplace.indexer.domain.models.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw.github_app_events;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.onlydust.marketplace.indexer.domain.models.raw.JsonDocument;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawInstallation;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawRepo;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
