+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/github_app_events/RawIssueCommentEvent.java
@@ -1,6 +1,7 @@
-package com.onlydust.marketplace.indexer.domain.models.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw.github_app_events;
 
 
+import com.onlydust.marketplace.indexer.domain.models.raw.*;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -10,10 +11,10 @@ import lombok.Value;
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(force = true)
 @ToString(callSuper = true)
-public class RawIssueEvent extends JsonDocument {
+public class RawIssueCommentEvent extends JsonDocument {
     String action;
     RawIssue issue;
+    RawComment comment;
     RawRepo repository;
     RawInstallation installation;
-    RawShortAccount assignee;
 }
