+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/github_app_events/RawIssueEvent.java
@@ -1,7 +1,7 @@
-package com.onlydust.marketplace.indexer.domain.models.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw.github_app_events;
 
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.onlydust.marketplace.indexer.domain.models.raw.*;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -11,10 +11,10 @@ import lombok.Value;
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(force = true)
 @ToString(callSuper = true)
-public class RawPullRequestEvent extends JsonDocument {
+public class RawIssueEvent extends JsonDocument {
     String action;
-    @JsonProperty("pull_request")
-    RawPullRequest pullRequest;
+    RawIssue issue;
     RawRepo repository;
     RawInstallation installation;
+    RawShortAccount assignee;
 }
