+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/models/raw/github_app_events/RawRepositoryEvent.java
@@ -1,6 +1,8 @@
-package com.onlydust.marketplace.indexer.domain.models.raw;
+package com.onlydust.marketplace.indexer.domain.models.raw.github_app_events;
 
 
+import com.onlydust.marketplace.indexer.domain.models.raw.JsonDocument;
+import com.onlydust.marketplace.indexer.domain.models.raw.RawRepo;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
