+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/ports/in/events/EventsInbox.java
@@ -1,6 +1,6 @@
 package com.onlydust.marketplace.indexer.domain.ports.in.events;
 
-import com.onlydust.marketplace.indexer.domain.models.raw.RawGithubAppEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawGithubAppEvent;
 
 public interface EventsInbox {
     void push(RawGithubAppEvent event);
