+++ domain/src/main/java/com/onlydust/marketplace/indexer/domain/services/events/RepositoryEventProcessorService.java
@@ -1,8 +1,8 @@
 package com.onlydust.marketplace.indexer.domain.services.events;
 
-import com.onlydust.marketplace.indexer.domain.models.clean.events.RepositoryEvent;
+import com.onlydust.marketplace.indexer.domain.models.clean.github_app_events.RepositoryEvent;
 import com.onlydust.marketplace.indexer.domain.models.exposition.GithubRepo;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawRepositoryEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawRepositoryEvent;
 import com.onlydust.marketplace.indexer.domain.ports.in.events.EventHandler;
 import com.onlydust.marketplace.indexer.domain.ports.in.indexers.RepoIndexer;
 import com.onlydust.marketplace.indexer.domain.ports.out.exposition.RepoStorage;
