+++ domain/src/test/java/com/onlydust/marketplace/indexer/domain/InstallationEventTest.java
@@ -2,7 +2,7 @@ package com.onlydust.marketplace.indexer.domain;
 
 import com.onlydust.marketplace.indexer.domain.models.RepoIndexingJobTrigger;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawAccount;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawInstallationEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawInstallationEvent;
 import com.onlydust.marketplace.indexer.domain.models.raw.RawRepo;
 import com.onlydust.marketplace.indexer.domain.ports.in.events.EventHandler;
 import com.onlydust.marketplace.indexer.domain.ports.out.jobs.RepoIndexingJobStorage;
@@ -43,6 +43,7 @@ public class InstallationEventTest {
         assertThat(events.get(0).getAccount().getId()).isEqualTo(onlyDust.getId());
         assertThat(events.get(0).getRepos()).hasSize(1);
         assertThat(events.get(0).getRepos().get(0).getId()).isEqualTo(marketplaceFrontend.getId());
-        verify(repoIndexingJobRepository).setInstallationForRepos(newInstallationEvent.getInstallation().getId(), new RepoIndexingJobTrigger(marketplaceFrontend.getId(), false, true));
+        verify(repoIndexingJobRepository).setInstallationForRepos(newInstallationEvent.getInstallation().getId(),
+                new RepoIndexingJobTrigger(marketplaceFrontend.getId(), false, true));
     }
 }
