+++ domain/src/test/java/com/onlydust/marketplace/indexer/domain/jobs/EventsInboxJobTest.java
@@ -1,8 +1,8 @@
 package com.onlydust.marketplace.indexer.domain.jobs;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawGithubAppEvent;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawInstallationEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawGithubAppEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawInstallationEvent;
 import com.onlydust.marketplace.indexer.domain.ports.in.events.EventHandler;
 import com.onlydust.marketplace.indexer.domain.ports.out.EventInboxStorage;
 import org.junit.jupiter.api.Test;
