+++ infrastructure/postgres-adapter/src/main/java/com/onlydust/marketplace/indexer/postgres/adapters/PostgresEventInboxStorage.java
@@ -1,7 +1,7 @@
 package com.onlydust.marketplace.indexer.postgres.adapters;
 
 import com.onlydust.marketplace.indexer.domain.exception.OnlyDustException;
-import com.onlydust.marketplace.indexer.domain.models.raw.RawGithubAppEvent;
+import com.onlydust.marketplace.indexer.domain.models.raw.github_app_events.RawGithubAppEvent;
 import com.onlydust.marketplace.indexer.domain.ports.out.EventInboxStorage;
 import com.onlydust.marketplace.indexer.postgres.entities.EventsInboxEntity;
 import com.onlydust.marketplace.indexer.postgres.repositories.raw.EventsInboxEntityRepository;
